requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==3.1
django-allauth==0.42.0
django-cors-headers==3.5.0
django-filter==2.3.0
django-rest-auth==0.9.5
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==2.0.1
oauthlib==3.1.0
packaging==20.4
Pillow==7.2.0
PyJWT==1.7.1
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10